my-property-solutions-master/my-property-solutions-master/pyproject.toml
[tool.poetry]
name = "my-property-solutions"
version = "0.0.1"
description = ""
authors = ["Stuart Dines <me@stuartdines.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aioredis = "=1.2.0"
asgiref = "=2.3.2"
async-timeout = "=2.0.1"
autobahn = "=18.12.1"
Automat = "=0.7.0"
billiard = "=3.5.0.5"
blessings = "=1.7"
bpython = "=0.18"
celery = "=4.2.1"
channels = "=2.1.6"
channels-redis = "=2.3.2"
chardet = "=3.0.4"
constantly = "=15.1.0"
curtsies = "=0.3.0"
daphne = "=2.2.4"
defusedxml = "=0.5.0"
Django = "=2.0"
django-allauth = "=0.36.0"
django-cleanup = "=2.1.0"
django-formtools = "=2.1"
django-widget-tweaks = "=1.4.2"
greenlet = "=0.4.15"
hiredis = "=0.3.1"
hyperlink = "=18.0.0"
incremental = "=17.5.0"
kombu = "=4.3.0"
oauthlib = "=2.1.0"
Pillow = "=5.3.0"
Pygments = "=2.3.1"
PyHamcrest = "=1.9.0"
pymemcache = "=2.1.1"
python-memcached = "=1.59"
python3-openid = "=3.1.0"
pytz = "=2018.5"
redis = "=3.0.1"
reportlab = "=3.5.21"
requests-oauthlib = "=1.0.0"
six = "=1.12.0"
sqlparse = "=0.3.0"
Twisted = "=18.9.0"
txaio = "=18.8.1"
typing = "=3.6.6"
vine = "=1.1.4"
wcwidth = "=0.1.7"
"zope.interface" = "=4.6.0"
amqp = "=2.5.1"
psycopg2 = "^2.8"
gunicorn = "^19.9"
sentry-sdk = "=0.13.4"
newrelic = "^5.4"

[tool.poetry.dev-dependencies]
black = "19.3b0"
flake8 = "3.7.8"
django-debug-toolbar = "1.11"
pre-commit = "1.18.3"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

